pyproject.toml
[build-system]
requires = ["setuptools>=45"]
build-backend = "setuptools.build_meta"

[project]
name = "fpdns"
version = "0.13.0"
description = "DNS fingerprinting tool"
requires-python = ">=3.8"
dependencies = ["dnspython>=2.0"]

[project.scripts]
fpdns = "fpdns.cli:main"

[tool.setuptools.packages.find]
include = ["fpdns*"]

[tool.pytest.ini_options]
testpaths = ["tests"]